feat(app structure): reorganize component structure and add new layout components
- Move existing components to a new layout directory for better organization.
- Add new components: About, ContactForm, CV, Footer, Navbar, ProjectCard, ProjectImage, and StackGrid.
- Update imports in page files to reflect new component paths.
