+++ app/cv/page.tsx
@@ -1,4 +1,4 @@
-import CV from '@/components/CV';
+import CV from '@/components/layout/CV';
 import { Button } from '@/components/ui/button';
 import { Download } from 'lucide-react';
